+++ doc/changelog.md
@@ -1,7 +1,14 @@
 ## Change Log
  
-### 2026-06-24
  
+### 2026-07-##
+- on app start or folder refresh, add an option in App Settings to auto select the prior Folder or the first Folder
+- add new cfg/settings_state.json
+- to the Folder list, add Forward, Backward, and Random buttons
+- add checkbox styling
+- release - [MyVideoExplorer-202607##-0.003 ](https://github.com/mbrowniebytes/MyVideoExplorer/releases/tag/202607##-0.003)
+- 
+### 2026-06-24
 - handle network shares
 - show loading media folder
 - auto select first folder
@@ -22,6 +29,5 @@
 - release - [MyVideoExplorer-20260624-0.002 ](https://github.com/mbrowniebytes/MyVideoExplorer/releases/tag/20260624-0.002)
 
 ### 2026-06-15
-
 - initial release
 - release - [MyVideoExplorer-20260615-0.001 ](https://github.com/mbrowniebytes/MyVideoExplorer/releases/tag/20260615-0.001)